pyproject.toml
[project]
name = "agentic-design-pattern"
version = "0.1.0"
description = ""
authors = [
    {name = "TAN PENG",email = "tanpeng8847@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk",
    "nest-asyncio (>=1.6.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
package-mode = false
